Only update tiddler on disk on a remote change
Debugging an issue where:

- User is editing `Foo.tid` locally on the file system
- User saves `Foo.tid` locally on the file system
- Save is propagated through MultiWikiAdapator and written back to disk
- User editing `Foo.tid` gets file system notification that the file has
  changed on disk

The intention is not to fix a file consistency problem,
the data saved back by `MultiWikiAdaptor` has been correct.

Instead, trying to fix a usability problem.
The local file system user has to manually reload the file after saving
it because the sync adapter also writes to the file afterwards.